+++ src/array.ts
@@ -10,6 +10,7 @@ export function array<T>(
       // Right value of "in" operator must be an object.
       // [MDN](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/in)
       typeof value === 'object' &&
+      value !== null &&
       // Check if value is ArrayLike or Iterable
       ('length' in value || Symbol.iterator in value)
         ? Array.from(value as ArrayLike<T> | Iterable<T>)